Move buffer stuff from pins.ts to buffer.ts
