+++ libs/core/pxt.json
@@ -31,6 +31,7 @@
         "dmac.cpp",
         "dmac.h",
         "timer.ts",
+        "buffer.ts",
         "pxtparts.json"
     ],
     "testFiles": [
